Remove more unnecessary indirection
